Revert "Drew/UI button edits in rest & gRPC, expanded hero banners for each section"
