+++ package.json
@@ -148,7 +148,6 @@
     "graphql-tools": "^6.1.0",
     "highland": "^2.13.5",
     "identity-obj-proxy": "^3.0.0",
-    "install": "^0.13.0",
     "js-beautify": "^1.14.0",
     "mali": "^0.21.0",
     "material-ui-dropzone": "^3.5.0",
@@ -156,9 +155,8 @@
     "ngrok": "^4.2.2",
     "node-fetch": "^2.6.0",
     "node-pre-gyp": "^0.6.39",
-    "node-sass": "npm:sass@^1.45.2",
+    "node-sass": "^4.14.1",
     "nodemon": "^1.19.4",
-    "npm": "^8.3.0",
     "npm-http-server": "^4.3.0",
     "path": "^0.12.7",
     "prop-types": "^15.6.2",
@@ -204,7 +202,7 @@
     "@testing-library/react": "^10.4.3",
     "@types/cookie": "^0.4.0",
     "@types/jest": "^26.0.4",
-    "@types/node": "^14.18.3",
+    "@types/node": "^14.0.20",
     "@types/node-fetch": "^2.5.7",
     "@types/react": "^16.9.41",
     "@types/react-dom": "^16.9.8",
@@ -212,6 +210,7 @@
     "@types/react-router-dom": "^5.1.6",
     "@types/set-cookie-parser": "0.0.6",
     "@types/uuid": "^8.3.1",
+    "@types/webrtc": "^0.0.19",
     "@types/ws": "^7.2.6",
     "@typescript-eslint/eslint-plugin": "^3.6.0",
     "@typescript-eslint/parser": "^3.10.1",
