+++ src/assets/style/base.scss
@@ -1,20 +1,18 @@
 
 * {
   box-sizing: border-box;
-  font-size: 12px;
   // background-color: lightgray;
-  // background-color: aquamarine;
 }
 
 // BULMA ADJUSTMENTS
 .hero {
-  height: 2.5rem; // nothing happened when changed
-  padding-top: 0.5rem; // makes purple banner change height
+  height: 2.5rem;
+  padding-top: 0.5rem;
 }
 
 .hero h3 {
   font-family: "Source Sans Pro", sans-serif;
-  font-weight: 600; // font boldness in banner (composer, workspace responses)
+  font-weight: 600;
 }
 
 input:focus {
