+++ src/assets/style/sidebar.scss
@@ -12,7 +12,6 @@
   border-radius: 2px;
   max-width: 26px;
   max-height: 26px;
-  padding-left: 2px;
 }
 
 .switch:focus {
@@ -20,20 +19,10 @@
   outline: none;
 }
 
-.add-header-gRPC-cookie-button {
-   color: $primary !important;
-  border-color: $primary !important;
-  width: 3px !important;
-  height: 3px !important;
-  margin-left: 2px !important;
-  margin-right: 2px !important;
-
-}
-
 .add-header-or-cookie-button {
   color: $primary !important;
   border-color: $primary !important;
-  min-width: 5px;
+  min-width: 80px;
 }
 
 .prettify-select {
@@ -102,7 +91,7 @@
   & :global(.react-codemirror2) {
          overflow: auto; /* you can use hidden here as well */
     }
-}
+ }
 
 .cm-s-responsebody .CodeMirror-wrap {
   height: 1000px !important;
