+++ src/assets/style/workspace.scss
@@ -4,10 +4,8 @@
 }
 
 .cards-dropdown {
-  border: 1px solid $neutral-200;
+  border: 1px solid $neutral-300;
   height: 24px;
-  border-radius: 4px;
-  margin-bottom: 5px;
 }
 
 .show-hide-tests {
