--- src/client/actions/actionTypes.ts
@@ -1,57 +1,0 @@
-// BUSINESS LOGIC ACTIONS
-export const REQRES_CLEAR = 'REQRES_CLEAR';
-export const REQRES_ADD = 'REQRES_ADD';
-export const REQRES_DELETE = 'REQRES_DELETE';
-export const REQRES_UPDATE = 'REQRES_UPDATE';
-export const SCHEDULED_REQRES_UPDATE = 'SCHEDULED_REQRES_UPDATE';
-export const SCHEDULED_REQRES_DELETE = 'SCHEDULED_REQRES_DELETE';
-export const UPDATE_GRAPH = 'UPDATE_GRAPH';
-
-export const CLEAR_GRAPH = 'CLEAR_GRAPH';
-export const CLEAR_ALL_GRAPH = 'CLEAR_ALL_GRAPH';
-
-export const GET_HISTORY = 'GET_HISTORY';
-export const DELETE_HISTORY = 'DELETE_HISTORY';
-export const CLEAR_HISTORY = 'CLEAR_HISTORY';
-
-export const GET_COLLECTIONS = 'GET_COLLECTIONS';
-export const DELETE_COLLECTION = 'DELETE_COLLECTION';
-export const COLLECTION_TO_REQRES = 'COLLECTION_TO_REQRES';
-export const COLLECTION_ADD = 'COLLECTION_ADD';
-export const COLLECTION_UPDATE = 'COLLECTION_UPDATE';
-
-export const SET_COMPOSER_WARNING_MESSAGE = 'SET_COMPOSER_WARNING_MESSAGE';
-
-export const RESET_COMPOSER_FIELDS = 'RESET_COMPOSER_FIELDS';
-
-export const SET_NEW_REQUEST_FIELDS = 'SET_NEW_REQUEST_FIELDS';
-export const SET_NEW_REQUEST_HEADERS = 'SET_NEW_REQUEST_HEADERS';
-export const SET_NEW_REQUEST_STREAMS = 'SET_NEW_REQUEST_STREAMS';
-export const SET_NEW_REQUEST_BODY = 'SET_NEW_REQUEST_BODY';
-export const SET_NEW_REQUEST_COOKIES = 'SET_NEW_REQUEST_COOKIES';
-export const SET_NEW_REQUEST_SSE = 'SET_NEW_REQUEST_SSE';
-export const SET_NEW_TEST_CONTENT = 'SET_NEW_TEST_CONTENT';
-
-export const SET_CURRENT_TAB = 'SET_CURRENT_TAB';
-
-export const SET_CHECKS_AND_MINIS = 'SET_CHECKS_AND_MINIS';
-
-export const SET_INTROSPECTION_DATA = 'SET_INTROSPECTION_DATA';
-
-export const SAVE_CURRENT_RESPONSE_DATA = 'SAVE_CURRENT_RESPONSE_DATA';
-
-// UI ACTIONS
-export const SET_COMPOSER_DISPLAY = 'SET_COMPOSER_DISPLAY';
-export const SET_SIDEBAR_ACTIVE_TAB = 'SET_SIDEBAR_ACTIVE_TAB';
-export const SET_WORKSPACE_ACTIVE_TAB = 'SET_WORKSPACE_ACTIVE_TAB';
-export const SET_RESPONSE_PANE_ACTIVE_TAB = 'SET_RESPONSE_PANE_ACTIVE_TAB';
-export const TOGGLE_DARK_MODE = 'TOGGLE_DARK_MODE';
-
-// OPENAPI ACTIONS
-export const SET_NEW_REQUESTS_OPENAPI = 'SET_NEW_REQUESTS_OPENAPI';
-export const SET_OPENAPI_SERVERS_GLOBAL = 'SET_OPENAPI_SERVERS_GLOBAL';
-export const SET_NEW_OPENAPI_SERVERS = 'SET_NEW_OPENAPI_SERVERS';
-export const SET_NEW_OPENAPI_PARAMETER = 'SET_NEW_OPENAPI_PARAMETER';
-export const SET_NEW_OPENAPI_REQUEST_BODY = 'SET_NEW_OPENAPI_REQUEST_BODY';
-
-export const QUEUE_OPENAPI_REQUESTS = 'QUEUE_OPENAPI_REQUESTS';
