+++ src/client/components/composer/ComposerContainer.jsx
@@ -250,7 +250,7 @@ const ComposerContainer = (props) => {
   const isDark = useSelector((store) => store.ui.isDark);
   
   return (
-    <div className="is-flex is-flex-direction-column">
+    <div className="is-flex is-flex-direction-column is-tall">
       {/* DROPDOWN PROTOCOL SELECTOR */}
       <NetworkDropdown
         onProtocolSelect={onProtocolSelect}
@@ -258,11 +258,8 @@ const ComposerContainer = (props) => {
         className="header-bar"
       />
 
-{/* is-not-7-5rem-tall pt-3 pl-3 pr-3 this was below */}
       {/* COMPOSER CONTENT ROUTING */}
-      <div className={`${isDark ? 'is-dark-400' : ''} add-vertical-scroll`}
-        style={{maxHeight: '30vh'}}>
-        
+      <div className={`${isDark ? 'is-dark-400' : ''} is-not-7-5rem-tall pt-3 pl-3 pr-3`}>
         {props.newRequestFields.network === 'rest' && (
           <RestContainer {...props} />
         )}
