+++ src/client/components/composer/NewRequest/ContentReqRowComposer.jsx
@@ -30,7 +30,6 @@ export default function ContentReqRowComposer({
         placeholder="Key"
         className="input"
         type="text"
-        style={{marginLeft: '3px'}}
         value={data.key}
         className={`${isDark ? 'is-dark-300' : ''} is-justify-content-center p-1 key`}
       />
@@ -39,10 +38,8 @@ export default function ContentReqRowComposer({
         placeholder="Value"
         className="input"
         type="text"
-        style={{marginLeft: '3px'}}
         value={data.value}
-        className={`${isDark ? 'is-dark-300' : ''} is-justify-content-center p-1 value`} 
-        // Removed is-flex-grow-4
+        className={`${isDark ? 'is-dark-300' : ''} is-justify-content-center is-flex-grow-4 p-1 value`}
       />
       <div className="is-flex is-justify-content-center is-align-items-center ml-1">
         <div className="delete m-auto" onClick={() => deleteItem(index)} />
