+++ src/client/components/composer/NewRequest/CookieEntryForm.jsx
@@ -87,7 +87,6 @@ class CookieEntryForm extends Component {
   }
 
   render() {
-    const addCookieName = '+';
     const cookiesArr = this.props.newRequestCookies.cookiesArr.map(
       (cookie, index) => (
         <ContentReqRowComposer
@@ -102,16 +101,14 @@ class CookieEntryForm extends Component {
     );
 
     return (
-      <div className="mt-2"
-      style={{margin: '10px'}}>
-        <div className="is-flex is-align-content-center">
+      <div className="mt-2">
+        <div className="is-flex is-justify-content-space-between is-align-content-center">
           <div className="composer-section-title">Cookies</div>
           <button
-            className={`${this.props.isDark ? 'is-dark-200' : ''} button add-header-gRPC-cookie-button`}
+            className={`${this.props.isDark ? 'is-dark-200' : ''} button is-small add-header-or-cookie-button`}
             onClick={() => this.addCookie(this.createDeepCookieCopy())}
-            style={{height: '2px', width: '2px'}}
           >
-            {addCookieName}
+            + Cookie
           </button>
         </div>
         <div>{cookiesArr}</div>
@@ -121,5 +118,3 @@ class CookieEntryForm extends Component {
 }
 
 export default CookieEntryForm;
-
-// is-justify-content-space-between was removed from + cookie button
