+++ src/client/components/composer/NewRequest/HeaderEntryForm.jsx
@@ -182,11 +182,11 @@ class HeaderEntryForm extends Component {
   
   render() {
     let headerName = 'Headers';
-    let addHeaderName = '+';
+    let addHeaderName = '+ Header';
     // let headerClass = 'composer_submit http'
     if (this.props.newRequestFields.gRPC) {
       headerName = 'Metadata';
-      addHeaderName = '+';
+      addHeaderName = '+ Metadata';
     }
 
     const headersArr = this.props.newRequestHeaders.headersArr.map(
@@ -203,13 +203,11 @@ class HeaderEntryForm extends Component {
     );
 
     return (
-      <div className="mt-2"
-      style={{margin: '10px'}}>
-        <div className="is-flex is-align-content-center">
+      <div className="mt-2">
+        <div className="is-flex is-justify-content-space-between is-align-content-center">
           <div className="composer-section-title">{headerName}</div>
           <button
-            className={`${this.props.isDark ? 'is-dark-200' : ''} button is-small add-header-gRPC-cookie-button`}
-            style={{height: '5px', width: '5px'}}
+            className={`${this.props.isDark ? 'is-dark-200' : ''} button is-small add-header-or-cookie-button`}
             onClick={() => this.addHeader()}
           >
             {addHeaderName}
@@ -222,7 +220,3 @@ class HeaderEntryForm extends Component {
 }
 
 export default HeaderEntryForm;
-
-
-
-// is-justify-content-space-between
