+++ src/client/components/composer/NewRequest/NewRequestButton.jsx
@@ -2,10 +2,9 @@ import React from 'react';
 
 const NewRequestButton = ({ onClick }) => (
   <button
-    className="button is-normal is-primary-100 add-request-button is-vertical-align-center is-justify-content-center no-border-please"
+    className="button is-normal is-fullwidth is-primary-100 is-button-footer is-margin-top-auto add-request-button is-vertical-align-center"
     onClick={onClick}
     type="button"
-    style={{margin: '10px'}}
   >
     Add New Request
   </button>
