+++ src/client/components/composer/NewRequest/WSEndpointEntryForm.jsx
@@ -28,8 +28,7 @@ const WSEndpointEntryForm = ({
   const isDark = useSelector((store) => store.ui.isDark);
 
   return (
-    <div className="is-flex is-justify-content-center"
-    style={{padding: '10px'}}>
+    <div className="is-flex is-justify-content-center">
       <div id="webSocketButton" className="no-border-please button is-ws">
         <span>WS</span>
       </div>
