+++ src/client/components/containers/App.tsx
@@ -2,11 +2,10 @@ import React, { useState, useEffect, JSXElementConstructor } from 'react';
 import { HashRouter } from 'react-router-dom';
 import ContentsContainer from './ContentsContainer';
 import SidebarContainer from './SidebarContainer';
-import ResponsePaneContainer from './ResponsePaneContainer';
-import RightSideContainer from './RightSideContainer';
 import historyController from '../../controllers/historyController';
 import collectionsController from '../../controllers/collectionsController';
 import UpdatePopUpContainer from './UpdatePopUpContainer';
+import ResponsePaneContainer from './ResponsePaneContainer';
 import { WindowExt } from '../../../types'
 import '../../../assets/style/App.scss';
 
@@ -23,28 +22,21 @@ const App = () => { //what type is being returned?
   }, []); // added the empty array in attempt to fix the issue of the app rerendering when the bargraph is clicked -Prince
 
   return (
-    // deleted is-tall below
-    <div className=" is-gapless is-tall">
+    <div className="is-gapless is-tall">
       <div
         id="app"
         className={`columns is-gapless ${!message && 'is-tall'} ${
           message && 'is-tall-message'
         }`}
       >
-        {/* <HashRouter> */}
-        {/* <div className="tile is-ancestor"> */}
-          {/* <div className="tile is-parent"> */}
-            <ContentsContainer />
-          {/* </div> */}
-          <div className="tile is-vertical add-vertical-scroll">
-            <SidebarContainer />
-            <ResponsePaneContainer />
-          </div>
-        {/* </div> */}
-        {/* </HashRouter> */}
+        <HashRouter>
+          <SidebarContainer />
+          <ContentsContainer />
+          <ResponsePaneContainer />
+        </HashRouter>
       </div>
       <UpdatePopUpContainer message={message} setMessage={setMessage} />
-  </div>
+    </div>
   );
 };
