+++ src/client/components/containers/ResponsePaneContainer.jsx
@@ -37,9 +37,7 @@ const ResponsePaneContainer = () => {
 
   return (
     <div
-      className={`${isDark ? 'is-dark-400' : 'is-divider-neutral-300'} box is-child`} // gigi playing with css
-      // className={`${isDark ? 'is-dark-400' : ''} column is-one-third is-flex is-flex-direction-column is-tall`}
-      style={{ height: '50vh', margin: "10px"}}
+      className={`${isDark ? 'is-dark-400' : ''} column is-one-third is-flex is-flex-direction-column is-tall`}
       id="responses"
     >
 
@@ -48,7 +46,7 @@ const ResponsePaneContainer = () => {
 
 
       {/* HEADER */}
-      <div className="hero is-primary is-flex is-flex-direction-row is-justify-content-center">
+      <div className="hero is-primary header-bar is-flex is-flex-direction-row is-justify-content-center">
         <ResponseTime currentResponse={currentResponse} />
         {currentResponse.responseSize && (
           <div className="response-size-placement">
