--- src/client/components/containers/RightSideContainer.jsx
@@ -1,15 +1,0 @@
-import React from React;
-import SidebarContainer from './SidebarContainer';
-import ResponsePaneContainer from './ResponsePaneContainer';
-
-const RightSideContainer = () => {
-    return (
-        <div classname="is-vertical">
-        	<SidebarContainer />
-          <ResponsePaneContainer />
-        </div>
-    )
-}
-
-
-export default RightSideContainer;
