+++ src/client/components/containers/SidebarContainer.tsx
@@ -28,26 +28,24 @@ const SidebarContainer = (props) => {
 
   return (
     <div
-      className={`box is-child ${isDark ? 'is-dark-400 add-vertical-scroll' : 'is-divider-neutral-300'}`} // gigi playing with css
-      // className={`column is-one-third is-flex is-flex-direction-column is-tall ${isDark ? 'is-dark-400' : ''}`}
-      style={{ height: '50vh', margin: "10px"}}
+      className={`column is-one-third is-flex is-flex-direction-column is-tall ${isDark ? 'is-dark-400' : ''}`}
       id="composer"
     >
 
       {/* HEADER */}
-      <div className="hero is-flex-direction-row is-primary has-text-centered">
+      <div className="hero is-flex-direction-row is-primary has-text-centered header-bar">
         {/* dark mode toggle goes here : https://www.w3schools.com/howto/howto_css_switch.asp */}
         <input
             id="darkModeSwitch"
             type="checkbox" 
-          className="switch mr-2 is-outlined is-warning"
+            className="switch mb-2 is-outlined is-warning"
             onChange={(e) => {
               handleDarkMode(e);
             }}
             checked={isDark}
         />
           <label htmlFor="darkModeSwitch"/>
-        <h3 className='has-text-justified align-self-center has-text-weight-bold'>Composer</h3>
+        <h3 className='has-text-justified align-self-center'>Composer</h3>
         <a>v1.0 </a>
       </div>
       <div className="tabs mb-0  header-bar">
