+++ src/client/reducers/index.ts
@@ -7,7 +7,6 @@ import uiReducer from './ui';
 // combine reducers
 const reducers = combineReducers({
   business: businessReducer,
-  // @ts-expect-error ts-migrate(2322) FIXME: Type '(state: { composerDisplay: string; sidebarAc... Remove this comment to see the full error message
   ui: uiReducer,
 });
