--- src/tsfixme.d.ts
@@ -1,2 +1,0 @@
-type $TSFixMe = any;
-type $TSFixMeFunction = (...args: any[]) => any;
