+++ tsconfig.json
@@ -25,8 +25,7 @@
     /* Advanced Options */
     "forceConsistentCasingInFileNames": true /* Disallow inconsistently-cased references to the same file. */
   },
-  // "include": ["./src/**/*.ts", "./src/**/*.tsx", "./main_process/*.ts", "src/assets/**/*.ts"],
-  "include": ["src/client/components/containers/ResponsePaneContainer.tsx"],
+  "include": ["./src/**/*.ts", "./src/**/*.tsx", "./main_process/*.ts", "src/assets/**/*.ts"],
   "exclude": ["node_modules", "dist", "build", "coverage", "mocks"],
   "typeRoots": ["./node_modules/@types"]
 }
